pyproject.toml
[project]
name = "preprocess-svs"
version = "0.1.0"
description = "preprocessing svs dataset"
readme = "README.md"
authors = [
    { name = "ccss17", email = "chansol0505@naver.com" }
]
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "g2pk2>=0.0.3",
    "hangul-dtw",
    "jamo>=0.4.1",
    "librosa>=0.11.0",
    "midii>=0.1.37",
    "numpy>=2.2.6",
    "openai-whisper>=20240930",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "praat-parselmouth>=0.4.6",
    "python-mecab-ko>=1.3.7",
    "pyworld>=0.3.5",
    "scipy>=1.15.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "ruff>=0.11.12",
]

[tool.uv.sources]
hangul-dtw = { git = "https://github.com/dirhq88/hangul_dtw.git" }

